+++ examples/inference/plot_sfh_recovery_test.py
@@ -37,10 +37,10 @@ obs = tengri.Observation(
 model = tengri.SEDModel.build(
     ssp,
     observation=obs,
-    sfh={"type": "dpl", "*": tengri.FREE},
+    sfh={"type": "dpl", "all_params": tengri.FREE},
     dust={
         "type": "two_component",
-        "*": tengri.FIXED,
+        "all_params": tengri.FIXED,
         "tau_diff": tengri.Uniform(0.0, 1.5),
         "slope": -0.7,
     },