+++ examples/spectroscopy/plot_bandheads_age_metallicity.py
@@ -145,14 +145,14 @@ model_solar = tengri.SEDModel.build(
     ssp,
     sfh={
         "type": "tsnorm",
-        "*": tengri.FIXED,
+        "all_params": tengri.FIXED,
         "peak_lbt_gyr": tengri.Uniform(0.03, 13.0),
         "width_gyr": 0.05,
         "log_total_mass": 10.0,
         "skew": 0.0,
         "trunc": 13.0,
     },
-    dust={"type": "two_component", "*": tengri.FIXED, "tau_diff": 0.0, "tau_bc": 0.0},
+    dust={"type": "two_component", "all_params": tengri.FIXED, "tau_diff": 0.0, "tau_bc": 0.0},
     redshift=Fixed(REDSHIFT),
     observation=obs,
 )
@@ -188,14 +188,19 @@ for j, age in enumerate(ages_multimet):
             ssp,
             sfh={
                 "type": "tsnorm",
-                "*": tengri.FIXED,
+                "all_params": tengri.FIXED,
                 "peak_lbt_gyr": age,
                 "width_gyr": 0.05,
                 "log_total_mass": 10.0,
                 "skew": 0.0,
                 "trunc": 13.0,
             },
-            dust={"type": "two_component", "*": tengri.FIXED, "tau_diff": 0.0, "tau_bc": 0.0},
+            dust={
+                "type": "two_component",
+                "all_params": tengri.FIXED,
+                "tau_diff": 0.0,
+                "tau_bc": 0.0,
+            },
             redshift=Fixed(REDSHIFT),
             observation=obs,
         )